+++ uplot/engine/PlotlyEngine5.py
@@ -1,10 +1,15 @@
 from uplot.interface import IPlotEngine, IFigure
+from uplot.default import DEFAULT
 
 
 class PlotlyEngine5(IPlotEngine):
+    # engine specific default parameters
     FILE_RESOLUTION_SCALE = 2
-    MARKER_SIZE = 8
     LINE_WIDTH = 2.5
+
+    @property
+    def name(self) -> str:
+        return 'plotly5'
 
     @classmethod
     def is_available(cls) -> bool:
@@ -13,10 +18,6 @@ class PlotlyEngine5(IPlotEngine):
             return True
         except ImportError:
             return False
-
-    @property
-    def figure_type(self) -> type:
-        return self.go.Figure
 
     @property
     def go(self):
@@ -30,11 +31,16 @@ class PlotlyEngine5(IPlotEngine):
     def __init__(self):
         import plotly.graph_objs as go
         import plotly.io as pio
+
         self._pio = pio
         self._go = go
+
         # load style
-        from uplot.engine.style.plotly import bmh
-        self._layout_style = bmh
+        if DEFAULT.style.lower() == 'bmh':
+            from uplot.engine.style.plotly import bmh
+            self._layout_style = bmh
+        else:
+            raise NotImplementedError(f'style not supported for plotly: {DEFAULT.style}')
 
     def figure(self, width: int, aspect_ratio: float) -> IFigure:
         from uplot.engine.PlotlyFigure5 import PlotlyFigure5