+++ uplot/engine/PlotlyFigure5.py
@@ -4,12 +4,13 @@ import numpy as np
 from numpy import ndarray
 from numpy.typing import ArrayLike
 
-import uplot.imtool as imtool
+import uplot.color as ucolor
+import uplot.utool as utool
 
-from uplot.interface import IFigure, LineStyle, MarkerStyle, AspectMode
+from uplot.interface import IFigure, LineStyle, MarkerStyle, AspectMode, Colormap
 from uplot.engine.PlotlyEngine5 import PlotlyEngine5
-from uplot.color import default_colors_list, decode_color, default_colors
-from uplot.routine import kwargs_extract
+from uplot.utool import Interpolator
+from uplot.default import DEFAULT
 
 
 class PlotlyFigure5(IFigure):
@@ -28,10 +29,11 @@ class PlotlyFigure5(IFigure):
 
     def __init__(self, engine: PlotlyEngine5):
         self._engine = engine
-        self._color_index = 0
+        self._color_scroller = ucolor.ColorScroller()
 
-        self._fig: engine.figure_type = engine.go.Figure()
+        self._fig: engine.go.Figure = engine.go.Figure()
         self._is_3d = None
+        self._colorbar_x_pos = 1.0
 
 
     def plot(self, x           : ArrayLike,
@@ -66,29 +68,29 @@ class PlotlyFigure5(IFigure):
             self._is_3d = False
 
         if marker_size is None:
-            marker_size = self.engine.MARKER_SIZE
+            # 1.33 - conversion gain between matplotlib and plotly marker size
+            marker_size = DEFAULT.marker_size*1.33
 
         if name is None:
             name = ''
             show_legend = False
         else:
-            show_legend = kwargs_extract(kwargs, name='showlegend', default=True)
+            show_legend = utool.kwargs_extract(kwargs, name='showlegend', default=True)
 
         if color is None:
-            color = self.current_color()
-            self.scroll_color()
+            color = self.scroll_color()
         elif not isinstance(color, str):
             # color specified for each point (x, y)
-            color = [ decode_color(c) for c in color ]
+            color = [ ucolor.name_to_hex(c) for c in color]
         else:
-            color = decode_color(color)
+            color = ucolor.name_to_hex(color)
 
         from uplot.engine.plotly import LINE_STYLE_MAPPING, MARKER_STYLE_MAPPING
         line_style = LINE_STYLE_MAPPING[line_style]
         marker_style = MARKER_STYLE_MAPPING[marker_style]
 
-        line = kwargs_extract(kwargs, name='line', default={})
-        marker = kwargs_extract(kwargs, name='marker', default={})
+        line = utool.kwargs_extract(kwargs, name='line', default={})
+        marker = utool.kwargs_extract(kwargs, name='marker', default={})
 
         mode = 'lines' if marker_style is None else 'lines+markers'
         if line_style == ' ': # no lines = scatter mode
@@ -105,7 +107,7 @@ class PlotlyFigure5(IFigure):
         marker['symbol'] = marker_style
         marker['size'] = marker_size
 
-        hoverlabel = kwargs_extract(kwargs, name='hoverlabel', default={})
+        hoverlabel = utool.kwargs_extract(kwargs, name='hoverlabel', default={})
         hoverlabel.setdefault('namelength', -1)
 
         if not self._is_3d:
@@ -147,16 +149,60 @@ class PlotlyFigure5(IFigure):
                   opacity=opacity,
                   **kwargs)
 
+    def surface3d(self, x            : ArrayLike,
+                        y            : ArrayLike,
+                        z            : ArrayLike,
+                        name         : str | None = None,
+                        show_colormap: bool = False,
+                        colormap     : Colormap = 'viridis',
+                        opacity      : float = 1.0,
+                        interpolation: Interpolator = 'cubic',
+                        interpolation_range: int = 100,
+                        **kwargs):
+        x = np.asarray(x)
+        y = np.asarray(y)
+        z = np.asarray(z)
+        assert x.ndim == y.ndim == 1, 'x, y must be 1D arrays'
+        assert z.ndim == 1 or z.ndim == 2, 'z must be 1D or 2D array'
+
+        if z.ndim == 2:
+            # uniform grid
+            assert (len(y), len(x)) == z.shape, 'uniform grid: x and y range must match z'
+            x, y = np.meshgrid(x, y)
+        else:
+            # non-uniform grid - array of points (x, y, z)
+            x, y, z = utool.array_to_grid(x, y, z,
+                                          interpolation=interpolation,
+                                          interpolation_range=interpolation_range)
+
+        self._is_3d = True
+
+        if show_colormap:
+            colorbar = dict(len=0.5, x=self._colorbar_x_pos)
+            self._colorbar_x_pos += 0.12
+        else:
+            colorbar = None
+
+        self._fig.add_surface(x=x, y=y, z=z,
+                              name=name,
+                              showlegend=(name != '') and (name is not None),
+                              showscale=show_colormap,
+                              colorscale=colormap,
+                              colorbar=colorbar,
+                              opacity=opacity,
+                              **kwargs)
+
+
     def imshow(self, image: ArrayLike, **kwargs):
         image = np.asarray(image)
-        value_range = imtool.estimate_range(image)
+        value_range = utool.image_range(image)
 
         self._is_3d = False
 
         self._fig.add_trace(self.engine.go.Image(
             z=image,
-            zmax=kwargs_extract(kwargs, name='zmax', default=[value_range]*4),
-            zmin=kwargs_extract(kwargs, name='zmin', default=[0]*4),
+            zmax=utool.kwargs_extract(kwargs, name='zmax', default=[value_range]*4),
+            zmin=utool.kwargs_extract(kwargs, name='zmin', default=[0]*4),
             **kwargs,
         ))
 
@@ -172,8 +218,9 @@ class PlotlyFigure5(IFigure):
     def legend(self, show: bool = True, **kwargs):
         self._fig.update_layout(legend=self.engine.go.layout.Legend(
             visible=show,
-            bgcolor=kwargs_extract(kwargs, name='bgcolor', default='rgba(255,255,255,0.8)'),
-            itemsizing=kwargs_extract(kwargs, name='itemsizing', default='constant'),
+            bgcolor=utool.kwargs_extract(kwargs, name='bgcolor', default='rgba(255,255,255,0.8)'),
+            itemsizing=utool.kwargs_extract(kwargs, name='itemsizing', default='constant'),
+            itemwidth=utool.kwargs_extract(kwargs, name='itemwidth', default=50),
             **kwargs,
         ))
 
@@ -249,15 +296,13 @@ class PlotlyFigure5(IFigure):
         self._fig.update_layout(scene=dict(zaxis=dict(range=[min_value, max_value])))
 
     def current_color(self) -> str:
-        color_name = default_colors_list[self._color_index]
-        return default_colors[color_name]
+        return self._color_scroller.current_color()
 
-    def scroll_color(self, count: int=1):
-        self._color_index += count
-        self._color_index %= len(default_colors_list)
+    def scroll_color(self, count: int=1) -> str:
+        return self._color_scroller.scroll_color(count)
 
     def reset_color(self):
-        self._color_index = 0
+        self._color_scroller.reset()
 
     def axis_aspect(self, mode: AspectMode):
         if self.is_3d:
@@ -281,11 +326,11 @@ class PlotlyFigure5(IFigure):
         image = image[..., :3] # RGBA -> RGB
         return image
 
-    def save(self, fname: str):
-        if '.html' in fname:
-            self._fig.write_html(fname)
+    def save(self, filename: str):
+        if '.html' in filename:
+            self._fig.write_html(filename)
         else:
-            self._fig.write_image(fname)
+            self._fig.write_image(filename)
 
     def close(self):
         self._fig = None