+++ uplot/interface.py
@@ -5,9 +5,8 @@ from typing import Protocol, runtime_checkable
 from abc import abstractmethod as abstract
 from numpy.typing import ArrayLike
 
-from uplot.LineStyle import LineStyle
-from uplot.MarkerStyle import MarkerStyle
-from uplot.AspectMode import AspectMode
+from uplot.utype import LineStyle, MarkerStyle, AspectMode, Colormap
+from uplot.utool import Interpolator
 
 # TODO: documentation
 
@@ -18,23 +17,25 @@ class IFigure(Protocol):
 
     Note
     ----
-    Probably it's not the best choice of the interface
-    but very common and familiar for MATLAB users.
+    This interface follows a Matplotlib-like interface for plotting, providing a familiar
+    environment for users, particularly those accustomed to MATLAB.
+    It may not be the best choice, but it offers common conventions and practices.
     """
+
     @property
     @abstract
     def engine(self) -> IPlotEngine:
-        return ...
+        ...
 
     @property
     @abstract
     def internal(self):
-        return ...
+        ...
 
     @property
     @abstract
     def is_3d(self) -> bool:
-        return ...
+        ...
 
     @abstract
     def plot(self, x           : ArrayLike,
@@ -59,6 +60,52 @@ class IFigure(Protocol):
                       marker_size : int | None = None,
                       opacity     : float = 1.0,
                       **kwargs):
+        ...
+
+    @abstract
+    def surface3d(self, x            : ArrayLike,
+                        y            : ArrayLike,
+                        z            : ArrayLike,
+                        name         : str | None = None,
+                        show_colormap: bool = False,
+                        colormap     : Colormap = 'viridis',
+                        opacity      : float = 1.0,
+                        interpolation: Interpolator = 'cubic',
+                        interpolation_range: int = 100,
+                        **kwargs):
+        """
+        Plot a surface in 3D space where the color scale corresponds to the z-values.
+        Two coordinate formats are supported:
+          - Uniform grid: x, y are 1D uniform ranges, z is a 2D array of corresponding values.
+          - Non-uniform grid (set of points): x, y, z are 1D arrays of corresponding points' coordinates in 3D space.
+            The non-uniform grid will be interpolated to a uniform grid using the specified **interpolator**.
+
+        Parameters
+        ----------
+        x, y, z : ArrayLike
+            Data values.
+
+        name : str or None, optional
+            The plot name, which will appear as the legend item.
+
+        show_colormap : bool, optional
+            Whether the colormap should be visualized as a bar alongside the plot.
+
+        colormap : Colormap, optional
+            A palette name string.
+
+        opacity : float, optional
+            Sets the opacity of the surface.
+
+        interpolation : Interpolator, optional
+            The interpolation method for the case when (x, y, z) is a non-uniform grid.
+
+        interpolation_range : int, optional
+            The number of points in the interpolated grid.
+
+        kwargs : dict
+            Other keyword arguments are forwarded to the underlying engine.
+        """
         ...
 
     @abstract
@@ -109,7 +156,7 @@ class IFigure(Protocol):
         ...
 
     @abstract
-    def scroll_color(self, count: int = 1):
+    def scroll_color(self, count: int = 1) -> str:
         ...
 
     @abstract
@@ -125,7 +172,7 @@ class IFigure(Protocol):
         ...
 
     @abstract
-    def save(self, fname: str):
+    def save(self, filename: str):
         ...
 
     @abstract
@@ -136,22 +183,23 @@ class IFigure(Protocol):
     def show(self, block: bool=True):
         ...
 
+
 @runtime_checkable
 class IPlotEngine(Protocol):
+
+    @property
+    @abstract
+    def name(self) -> str:
+        ...
 
     @classmethod
     @abstract
     def is_available(cls) -> bool:
         ...
 
-    @property
-    @abstract
-    def figure_type(self) -> type:
-        return ...
-
     @abstract
     def figure(self, width: int, aspect_ratio: float) -> IFigure:
         """
-        Factory method for a figure creation
+        Factory method for the figure creation
         """
         ...