+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## `[v0.3.0]` - 06.12.2023
+
+#### Added
+* `[interface]` singleton `DEFAULT` for storing and controlling default parameters.
+* `[interface]` 3d surface plotting: `figure.surface3d(...)`
+* `[color]` class `ColorScroller` for maintaining automatic color switching for plotting.
+
+#### Changed
+* `[engine]` the engine management system reworked: `engine.get()`, `engine.available()`, `engine.register()`.
+
+#### Fixed
+* `[engine.plotly5]` legend item width increased to show dashed line correctly.
+* `[engine.matplot]` problem with `figure.close()` fixed.
+
+
 ## `[v0.2.1]` - 13.11.2023
 
 #### Fixed